Fix pedantic warnings on MSVC, GCC, and Clang

+++ examples/appearance/multiplot/colororder/colororder_2.cpp
@@ -10,7 +10,7 @@ int main() {
 
     hold(on);
     for (size_t r = 1; r <= 7; ++r) {
-        auto x = linspace(0, r, 500);
+        auto x = linspace(0., static_cast<double>(r), 500);
         auto y =
             transform(x, [&](double x) { return sqrt(pow(r, 2) - pow(x, 2)); });
         plot(x, y)->line_width(15);
