+++ examples/geography/geoplot/geoplot_6.cpp
@@ -121,7 +121,7 @@ void eurotrip_solver::draw() {
 
     vector<double> sorted_lat;
     vector<double> sorted_lon;
-    for (const int &idx : best_tour_) {
+    for (const size_t &idx : best_tour_) {
         sorted_lat.emplace_back(lat_[idx]);
         sorted_lon.emplace_back(lon_[idx]);
     }
