+++ examples/graphs/graph/graph_5.cpp
@@ -9,7 +9,7 @@ int main() {
     auto g = graph(edges);
     g->layout_algorithm(network::layout::force);
 
-    for (size_t i = 0; i < 300; ++i) {
+    for (int i = 0; i < 300; ++i) {
         g->layout_iterations(i);
         std::this_thread::sleep_for(std::chrono::milliseconds(100));
     }
